# Conflicts:
#	app/src/main/java/com/helplive/bcm208assignment/data/DatabaseHandler.java

+++ app/src/main/java/com/helplive/bcm208assignment/RegisterApplicant.java
@@ -8,6 +8,8 @@ import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.helplive.bcm208assignment.model.Applicant;
+
 public class RegisterApplicant extends AppCompatActivity {
 
     @Override
@@ -30,7 +32,7 @@ public class RegisterApplicant extends AppCompatActivity {
         String confpassword = editTextPass2.getText().toString().trim();
         String email = editTextEmail.getText().toString().trim();
         String monthlySalaryStr = editTextMonthSal.getText().toString().trim();
-        Integer monthlySalary = 0;
+        Double monthlySalary = 0.0;
 
         //Toast.makeText(this.getBaseContext(),"TEST",Toast.LENGTH_SHORT).show();
 
@@ -61,7 +63,7 @@ public class RegisterApplicant extends AppCompatActivity {
         }
 
         try{
-           Integer.parseInt(monthlySalaryStr);
+           Double.parseDouble(monthlySalaryStr);
         }catch (Exception e){
             Toast.makeText(this.getBaseContext(),"Please fill in the monthly salary field",Toast.LENGTH_SHORT).show();
             editTextMonthSal.requestFocus();
@@ -87,7 +89,7 @@ public class RegisterApplicant extends AppCompatActivity {
             return;
         }
         //4. Monthly Salary can't be 0
-        if(monthlySalary <= 0){
+        if(monthlySalary <= 0.0){
             Toast.makeText(this.getBaseContext(),"Monthly salary cannot be RM 0",Toast.LENGTH_SHORT).show();
             editTextEmail.requestFocus();
             return;
